% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp_functions.R
\name{ellipse_sort}
\alias{ellipse_sort}
\title{Ellipse Sort}
\usage{
ellipse_sort(data)
}
\arguments{
\item{data}{A square numeric proximity matrix (either n × n or p × p), representing pairwise distances or similarities between items.}
}
\value{
An integer vector representing the reordered indices of the matrix rows.
}
\description{
This function applies an Rank-2-Ellipse seriation method to reorder a proximity matrix.
}
