% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate-v.R
\name{estimate_v}
\alias{estimate_v}
\title{Estimates the effect size V and its standard error}
\usage{
estimate_v(data, g1, g2)
}
\arguments{
\item{data}{The dataset used to estimate V}

\item{g1}{count data the first group in each of $n$ categories}

\item{g2}{count data the second group in each of $n$ categories}
}
\description{
Uses maximum likelihood to estimate the ROC curve, estimates
the area under the curve, and returns the AUC, V and standard
error of V
}
