% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_gapr.R
\name{fit}
\alias{fit}
\title{Fit statistical models to gapminder data}
\usage{
fit(obj, effects = c("fixed", "mixed"))
}
\arguments{
\item{obj}{An object of class \code{"gapr"} generated by \code{\link{load_gapr}}.}

\item{effects}{Specify what kind of effects to be used in the model.}
}
\value{
An object of class \code{"gapr_fit"} which includes the model details
as well as the data set and the kind of \code{effects} used.
}
\description{
Fits linear regression models with fixed and random effects to the gapminder data.
}
\note{
A dedicated \code{fit.gapr} method is provided
for objects of class \code{"gapr_fit"}
.
}
\examples{
dat <- load_gapr("country", "Ireland")
mod1 <- fit(dat)
mod2 <- fit(dat, effects = "mixed")
plot(mod1)
plot(mod2)
}
\seealso{
\code{\link{load_gapr}}
}
\author{
Michael Doherty - <\email{michael.doherty.2019@mumail.ie}>
}
