% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{GAP_Blue_White_Red}
\alias{GAP_Blue_White_Red}
\title{Color palette: GAP_Blue_White_Red}
\format{
An object of class \code{character} of length 20.
}
\usage{
GAP_Blue_White_Red
}
\description{
A diverging color palette from blue to white to red, suitable for visualizing correlations.
}
\keyword{datasets}
