% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{GAP_d}
\alias{GAP_d}
\title{Color palette: GAP_d}
\format{
An object of class \code{character} of length 16.
}
\usage{
GAP_d
}
\description{
A 16-category discrete color palette.
}
\keyword{datasets}
