% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{GAP_Rainbow}
\alias{GAP_Rainbow}
\title{Color palette: GAP_Rainbow}
\format{
An object of class \code{character} of length 13.
}
\usage{
GAP_Rainbow
}
\description{
A color palette with 13 colors in rainbow used for continuous data.
}
\keyword{datasets}
