% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{grayscale_palette}
\alias{grayscale_palette}
\title{Color palette: grayscale_palette}
\format{
An object of class \code{character} of length 2.
}
\usage{
grayscale_palette
}
\description{
A simple two-color grayscale palette from white to black, often used for binary data visualization.
}
\keyword{datasets}
