% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_gapr.R
\name{load_gapr}
\alias{load_gapr}
\title{Gapminder Data}
\usage{
load_gapr(
  summarize = c("none", "year", "country", "continent"),
  value = c("none")
)
}
\arguments{
\item{summarize}{Variable on which the gapminder dataset will be summarized.}

\item{value}{Value(s) to condition on when filtering the gapminder dataset.}
}
\value{
A data.frame of 6 columns and 1704 observations on different countries
from year 1952 to 2007
}
\description{
Loads in gapminder dataset from Github.
}
\note{
A dedicated \code{\link{fit}} and \code{\link{plot_gapr}} functions are
provided for objects of class \code{"gapr"}.
}
\examples{
dat <- load_gapr("country", "Ireland")
dat <- load_gapr("continent", "Europe")
dat <- load_gapr("year")

}
\seealso{
\code{\link{fit}}, \code{\link{plot_gapr}}
}
\author{
Syed Baryalay - <\email{syed.baryalay.2020@mumail.ie"}>
}
