% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gapr.R
\name{plot.gapr}
\alias{plot.gapr}
\title{Plot gapr output}
\usage{
\method{plot}{gapr}(obj, var = c("country", "continent"), val, ...)
}
\arguments{
\item{obj}{An object of class \code{"gapr"} outputted from the \code{\link{load_gapr}} function.}
}
\value{
A ggplot plot which plots \code{lifeExp} against \code{year}
}
\description{
Plot \code{"lifeExp"} over the years, of N countries or continents
}
\examples{
dat <- load_gapr()
plot(dat, continent, c("Asia", "Europe", "Africa"))
plot(dat, country, c("China", "Ireland", "Nigeria"))
}
\seealso{
\code{\link{load_gapr}}
}
\author{
Syed Baryalay - <\email{syed.baryalay.2020@mumail.ie}>
}
