% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gapr_fit.R
\name{plot.gapr_fit}
\alias{plot.gapr_fit}
\title{Plot gapr fit output}
\usage{
\method{plot}{gapr_fit}(model_fit)
}
\arguments{
\item{model_fit}{An object of class \code{"gapr_fit"} outputted from the \code{link{fit}} function.}
}
\value{
A ggplot plot which plots \code{lifeExp} against \code{year}.
}
\description{
Plot \code{"lifeExp"} over the years, worldwide, by country or by continent.
}
\examples{
dat <- load_gapr("country", "Ireland")
mod1 <- fit(dat)
mod2 <- fit(dat, effects = "mixed")
plot(mod1)
plot(mod2)
}
\seealso{
\code{\link{fit}}
}
\author{
Syed Baryalay - <\email{syed.baryalay.2020@mumail.ie}>
}
