% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gapr_fit.R
\name{plot}
\alias{plot}
\title{Plot gapr fit output}
\usage{
plot(x, ...)
}
\arguments{
\item{x}{An object of class \code{"gapr_fit"} outputted from the \code{\link{fit}} function.}

\item{...}{Optional arguments that can be passed to the plot method.}
}
\value{
A ggplot plot which plots \code{lifeExp} against \code{year}.
}
\description{
Plot \code{"lifeExp"} over the years, worldwide, by country or by continent.
}
\examples{
dat <- load_gapr("country", "Ireland")
mod1 <- fit(dat)
mod2 <- fit(dat, effects = "mixed")
plot(mod1)
plot(mod2)
}
\seealso{
\code{\link{load_gapr}} \code{\link{fit}}
}
\author{
Syed Baryalay - <\email{syed.baryalay.2020@mumail.ie}>
}
