% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gapr.R
\name{plot_gapr}
\alias{plot_gapr}
\title{Plot gapr output}
\usage{
plot_gapr(x, var = continent, val = c("Asia", "Europe"))
}
\arguments{
\item{x}{An object of class \code{"gapr"} outputted from the \code{\link{load_gapr}} function.}

\item{var}{A data variable from the gapminder on which the data is summarized.}

\item{val}{A value for \code{country} or \code{continent} data variable which need to be summarized.}
}
\value{
A ggplot plot which plots \code{lifeExp} against \code{year}
}
\description{
Plot \code{"lifeExp"} over the years, of N countries or continents
}
\examples{
dat <- load_gapr()
plot_gapr(dat, continent, c("Asia", "Europe", "Africa"))
plot_gapr(dat, country, c("China", "Ireland", "Nigeria"))
}
\seealso{
\code{\link{load_gapr}}
}
\author{
Syed Baryalay - <\email{syed.baryalay.2020@mumail.ie}>
}
