% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_functions.R
\name{.compute_gar_core}
\alias{.compute_gar_core}
\title{Internal function to compute Anti-Robinson or GAR/ RGAR score}
\usage{
.compute_gar_core(mat_sorted, w = NULL, normalize = FALSE)
}
\arguments{
\item{mat_sorted}{A numeric, symmetric sorted distance matrix.}

\item{w}{An integer window size (if NULL, evaluates all triplets globally).}

\item{normalize}{Logical. If TRUE, returns a proportion; otherwise returns the raw count.}
}
\value{
A numeric value indicating the number (or proportion) of Anti-Robinson violations.
}
\description{
This function is not exported. It serves as the core computational engine
for evaluating Anti-Robinson (AR), Generalized Anti-Robinson (GAR), and
Relative GAR (RGAR) scores.
}
\keyword{internal}
