% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gapr.R, R/plot_gapr_fit.R
\name{plot}
\alias{plot}
\title{Plot gapr output}
\usage{
plot(x, ...)

plot(x, ...)
}
\arguments{
\item{x}{An object of class \code{"gapr_fit"} outputted from the \code{\link{fit}} function.}

\item{...}{Optional arguments that can be passed to the plot method.}

\item{var}{A data variable from the gapminder on which the data is summarized.}

\item{val}{A value for \code{country} or \code{continent} data variable which need to be summarized.}
}
\value{
A ggplot plot which plots \code{lifeExp} against \code{year}

A ggplot plot which plots \code{lifeExp} against \code{year}.
}
\description{
Plot \code{"lifeExp"} over the years, of N countries or continents

Plot \code{"lifeExp"} over the years, worldwide, by country or by continent.
}
\examples{
dat <- load_gapr()
plot(dat, continent, c("Asia", "Europe", "Africa"))
plot(dat, country, c("China", "Ireland", "Nigeria"))
dat <- load_gapr("country", "Ireland")
mod1 <- fit(dat)
mod2 <- fit(dat, effects = "mixed")
plot(mod1)
plot(mod2)
}
\seealso{
\code{\link{load_gapr}}

\code{\link{fit_gapr}} \code{\link{fit}}
}
\author{
Syed Baryalay - <\email{syed.baryalay.2020@mumail.ie}>
}
